refactor: consolidate aave utils
