+++ src/components/ui/lending/BorrowComponent.tsx
@@ -6,17 +6,19 @@ import {
   AccordionTrigger,
 } from "@/components/ui/Accordion";
 import { ScrollBoxSupplyBorrowAssets } from "@/components/ui/lending/ScrollBoxSupplyBorrowAssets";
-import { useAaveChain, useIsWalletTypeConnected } from "@/store/web3Store";
+import useWeb3Store, {
+  useAaveChain,
+  useIsWalletTypeConnected,
+} from "@/store/web3Store";
+import { useAaveFetch } from "@/utils/aave/fetch";
+import BorrowUnownedCard from "@/components/ui/lending/BorrowUnownedCard";
+import BorrowOwnedCard from "@/components/ui/lending/BorrowOwnedCard";
+import { WalletType } from "@/types/web3";
 import {
   AaveReserveData,
   AaveReservesResult,
   UserBorrowPosition,
-  useAaveFetch,
-} from "@/utils/aave/fetch";
-import BorrowUnownedCard from "@/components/ui/lending/BorrowUnownedCard";
-import BorrowOwnedCard from "@/components/ui/lending/BorrowOwnedCard";
-import { WalletType } from "@/types/web3";
-import useWeb3Store from "@/store/web3Store";
+} from "@/types/aave";
 
 const BorrowComponent: React.FC = () => {
   const [borrowableReserves, setBorrowableReserves] = useState<
