+++ src/components/ui/lending/BorrowOwnedCard.tsx
@@ -13,11 +13,10 @@ import {
   CardDescription,
 } from "@/components/ui/Card";
 import type { Token, Chain } from "@/types/web3";
-import { UserBorrowPosition } from "@/utils/aave/fetch";
+import { UserBorrowPosition, RateMode } from "@/types/aave";
 import { formatBalance } from "@/utils/formatters";
 import { getChainByChainId } from "@/config/chains";
 import RepayModal from "@/components/ui/lending/RepayModal";
-import { RateMode } from "@/utils/aave/interact";
 import AssetDetailsModal from "@/components/ui/lending/AssetDetailsModal";
 
 interface BorrowOwnedCardProps {
